// frontend/web/src/services/API/proserve-wb-packaging-api/models/Pipeline.ts
export interface Pipeline {
    /**
     * Pipeline ID
     * @type {string}
     * @memberof Pipeline
     */
    pipelineId: string;
    /**
     * The instance types of the infrastructure configuration. You can specify one or more instance types to use for this build
     * @type {Array<string>}
     * @memberof Pipeline
     */
    buildInstanceTypes: Array<string>;
    /**
     * The description of the pipeline
     * @type {string}
     * @memberof Pipeline
     */
    pipelineDescription: string;
    /**
     * The name of the pipeline
     * @type {string}
     * @memberof Pipeline
     */
    pipelineName: string;
    /**
     * The schedule of the pipeline. See https://docs.aws.amazon.com/imagebuilder/latest/userguide/cron-expressions.html
     * @type {string}
     * @memberof Pipeline
     */
    pipelineSchedule: string;
    /**
     * Recipe ID of the recipe associated to the pipeline
     * @type {string}
     * @memberof Pipeline
     */
    recipeId: string;
    /**
     * Recipe Name of the recipe associated to the pipeline
     * @type {string}
     * @memberof Pipeline
     */
    recipeName: string;
    /**
     * Recipe version ID of the recipe version associated to the pipeline
     * @type {string}
     * @memberof Pipeline
     */
    recipeVersionId: string;
    /**
     * Recipe version Name of the recipe version associated to the pipeline
     * @type {string}
     * @memberof Pipeline
     */
    recipeVersionName: string;
    /**
     * Product ID associated with the pipeline for automatic version creation
     * @type {string}
     * @memberof Pipeline
     */
    productId?: string;
    /**
     * Status of the pipeline (CREATING/CREATED/UPDATING/RELEASED/RETIRED/FAILED)
     * @type {string}
     * @memberof Pipeline
     */
    status: string;
    /**
     * Create Date
     * @type {string}
     * @memberof Pipeline
     */
    createDate: string;
    /**
     * The Amazon Resource Name (ARN) of the distribution configuration
     * @type {string}
     * @memberof Pipeline
     */
    distributionConfigArn?: string;
    /**
     * The Amazon Resource Name (ARN) of the infrastructure configuration
     * @type {string}
     * @memberof Pipeline
     */
    infrastructureConfigArn?: string;
    /**
     * The Amazon Resource Name (ARN) of the pipeline
     * @type {string}
     * @memberof Pipeline
     */
    pipelineArn?: string;
    /**
     * Created By
     * @type {string}
     * @memberof Pipeline
     */
    createdBy: string;
    /**
     * Last Update Date
     * @type {string}
     * @memberof Pipeline
     */
    lastUpdateDate: string;
    /**
     * Last Update By
     * @type {string}
     * @memberof Pipeline
     */
    lastUpdatedBy: string;
}